# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Inquira",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "V1 Authentication"
        ],
        "summary": "Register User",
        "description": "Register a new user and immediately issue login session cookie.",
        "operationId": "register_user_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "V1 Authentication"
        ],
        "summary": "Login User",
        "description": "Login user and issue session cookie.",
        "operationId": "login_user_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/preferences": {
      "get": {
        "tags": [
          "V1 Preferences"
        ],
        "summary": "Get Preferences",
        "operationId": "get_preferences_api_v1_preferences_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "V1 Preferences"
        ],
        "summary": "Update Preferences",
        "operationId": "update_preferences_api_v1_preferences_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/preferences/api-key": {
      "put": {
        "tags": [
          "V1 Preferences"
        ],
        "summary": "Set Api Key",
        "operationId": "set_api_key_api_v1_preferences_api_key_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "V1 Preferences"
        ],
        "summary": "Delete Api Key",
        "operationId": "delete_api_key_api_v1_preferences_api_key_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/reset": {
      "post": {
        "tags": [
          "V1 Admin"
        ],
        "summary": "Reset Everything",
        "description": "One-time destructive reset for legacy/users/schema data.\n\nRequires:\n  - INQUIRA_ENABLE_RESET=1\n  - token matching INQUIRA_RESET_TOKEN",
        "operationId": "reset_everything_api_v1_admin_reset_post",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/test-gemini": {
      "post": {
        "tags": [
          "V1 Admin"
        ],
        "summary": "Test Gemini Api Key",
        "description": "Validate a Gemini key by issuing a lightweight test request.",
        "operationId": "test_gemini_api_key_api_v1_admin_test_gemini_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeminiTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiKeyUpdateRequest": {
        "properties": {
          "api_key": {
            "type": "string",
            "minLength": 1,
            "title": "Api Key"
          }
        },
        "type": "object",
        "required": [
          "api_key"
        ],
        "title": "ApiKeyUpdateRequest"
      },
      "AuthUserResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "plan": {
            "type": "string",
            "title": "Plan"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "username",
          "plan"
        ],
        "title": "AuthUserResponse",
        "description": "Authenticated user metadata returned to frontend."
      },
...
      "GeminiTestRequest": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "description": "Google Gemini API key to test"
          }
        },
        "type": "object",
        "required": [
          "api_key"
        ],
        "title": "GeminiTestRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 6,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "description": "Login payload."
      },
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "description": "Simple message response."
      },
      "PreferencesResponse": {
        "properties": {
          "selected_model": {
            "type": "string",
            "title": "Selected Model",
            "default": "gemini-2.5-flash"
          },
          "schema_context": {
            "type": "string",
            "title": "Schema Context",
            "default": ""
          },
          "allow_schema_sample_values": {
            "type": "boolean",
            "title": "Allow Schema Sample Values",
            "default": false
          },
          "chat_overlay_width": {
            "type": "number",
            "title": "Chat Overlay Width",
            "default": 0.25
          },
          "is_sidebar_collapsed": {
            "type": "boolean",
            "title": "Is Sidebar Collapsed",
            "default": true
          },
          "hide_shortcuts_modal": {
            "type": "boolean",
            "title": "Hide Shortcuts Modal",
            "default": false
          },
          "active_workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Workspace Id"
          },
          "active_dataset_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Dataset Path"
          },
          "active_table_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Table Name"
          },
          "api_key_present": {
            "type": "boolean",
            "title": "Api Key Present",
            "default": false
          }
        },
        "type": "object",
        "title": "PreferencesResponse"
      },
      "PreferencesUpdateRequest": {
        "properties": {
          "selected_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Model"
          },
          "schema_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Context"
          },
          "allow_schema_sample_values": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allow Schema Sample Values"
          },
          "chat_overlay_width": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 0.9,
                "minimum": 0.1
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Overlay Width"
          },
          "is_sidebar_collapsed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Sidebar Collapsed"
          },
          "hide_shortcuts_modal": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hide Shortcuts Modal"
          },
          "active_workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Workspace Id"
          },
          "active_dataset_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Dataset Path"
          },
          "active_table_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Table Name"
          }
        },
        "type": "object",
        "title": "PreferencesUpdateRequest"
      },
      "RegisterRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 6,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "RegisterRequest",
        "description": "Register payload."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}